backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic-settings==2.5.2
httpx==0.27.2
SQLAlchemy==2.0.36
alembic==1.13.3
psycopg[binary]==3.2.3
pgvector==0.2.5
python-multipart==0.0.9
sentence-transformers==3.2.1
pytest==8.3.3
pytest-asyncio==0.24.0
typer==0.12.5
beautifulsoup4==4.12.3
gspread
